UahDataset: Use name of image and not of xml on the image dicts.
Previously we had a mapping of the location of the xml
(for example ~/dataset/gba_dataset/2016_video001/xml/000121.xml -> [0, 1]).
We want the images, not the xmls, so read the full path of the file from
the xml and use it to build the dict.

Sadly, this means that we must regenerate the xml, when we download the
dataset. As a future improvement we may want to construct the path
ourselfs using only the image name (which is in filename instead of in
path).

This should be doable as we control the folder structure of the dataset
and the folders of the frames is ./../FRAMES/ with respect to the folder
of the xml.