Implement reading of dataset and division of train, query and gallery.
We don't save the split as it does a new one in a short amount of time
(30 seconds), so it make sense to regenerate them every time we
instantiate the class.

Also in ilids only consider two images of every class for query and
gallery, but here we consider them all, by poping two random element
of a set, one for query, one for gallery of the testing data, until we
only have zero or one images left.