Introduce the concept of valid_(yolo|iou)_results.
YOLO may not always detect the correct results that are annotated, so
the iteration through the results of `yolo_results` (after iou) or
`yolo_ids` (in the `calculate_best_fit_yolo`) may fail because we are
trying to iterate an empty list.

To avoid that error condition gate the iteration after a boolean
that is true if the length of the lists is greater than 0.